chore(dev): restructure backend
