+++ src/bp/core/botpress.ts
@@ -1,6 +1,9 @@
 import * as sdk from 'botpress/sdk'
 import lang from 'common/lang'
+import { CMSService } from 'core/cms'
+import { BotpressConfig, ConfigProvider } from 'core/config'
 import { buildUserKey, converseApiEvents } from 'core/converse'
+import { SessionIdFactory } from 'core/dialog'
 import { addStepToEvent, EventCollector, StepScopes, StepStatus, EventEngine } from 'core/events'
 import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } from 'core/logger'
 import { copyDir } from 'core/misc/pkg-fs'
@@ -20,9 +23,9 @@ import { startLocalActionServer } from '../cluster'
 import { setDebugScopes } from '../debug'
 
 import { createForGlobalHooks } from './api'
-import { BotpressConfig } from './config/botpress.config'
-import { ConfigProvider } from './config/config-loader'
+
 import Database from './database'
+import { StateManager } from './dialog/state-manager'
 import { ModuleLoader } from './module-loader'
 import { WellKnownFlags } from './sdk/enums'
 import { Event } from './sdk/impl'
@@ -33,14 +36,11 @@ import { AlertingService } from './services/alerting-service'
 import AuthService from './services/auth/auth-service'
 import { BotMonitoringService } from './services/bot-monitoring-service'
 import { BotService } from './services/bot-service'
-import { CMSService } from './services/cms'
 import { DecisionEngine } from './services/dialog/decision-engine'
 import { DialogEngine } from './services/dialog/dialog-engine'
 import { DialogJanitor } from './services/dialog/janitor'
-import { SessionIdFactory } from './services/dialog/session/id-factory'
 import { HintsService } from './services/hints'
 import { Hooks, HookService } from './services/hook/hook-service'
-import { StateManager } from './services/middleware/state-manager'
 import { MigrationService } from './services/migration'
 import { MonitoringService } from './services/monitoring'
 import { NotificationsService } from './services/notification/service'
