+++ src/bp/core/database/tables/bot-specific/index.ts
@@ -1,4 +1,3 @@
-export * from './dialog_sessions'
 export * from './ghost_content'
 export * from './notifications'
 export * from './conversations'
