+++ src/bp/core/dialog/sessions/dialog_sessions-table.ts
