+++ src/bp/core/events/event-collector.ts
@@ -1,7 +1,7 @@
 import * as sdk from 'botpress/sdk'
 import { ConfigProvider } from 'core/config/config-loader'
 import Database from 'core/database'
-import { SessionIdFactory } from 'core/services/dialog/session/id-factory'
+import { SessionIdFactory } from 'core/dialog'
 import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
 import Knex from 'knex'
