+++ src/bp/core/routers/bots/index.ts
@@ -3,15 +3,14 @@ import { Serialize } from 'cerialize'
 import { decodeFolderPath, UnexpectedError } from 'common/http'
 import { gaId, machineUUID } from 'common/stats'
 import { FlowView } from 'common/typings'
-import { BotpressConfig } from 'core/config/botpress.config'
-import { ConfigProvider } from 'core/config/config-loader'
+import { CMSService } from 'core/cms'
+import { BotpressConfig, ConfigProvider } from 'core/config'
 import { LogsService } from 'core/logger'
 import { ModuleLoader } from 'core/module-loader'
 import ActionServersService from 'core/services/action/action-servers-service'
 import ActionService from 'core/services/action/action-service'
 import AuthService, { TOKEN_AUDIENCE } from 'core/services/auth/auth-service'
 import { BotService } from 'core/services/bot-service'
-import { CMSService } from 'core/services/cms'
 import { FlowService, MutexError, TopicSchema } from 'core/services/dialog/flow/service'
 import { MediaServiceProvider } from 'core/services/media'
 import { NotificationsService } from 'core/services/notification/service'
