+++ src/bp/core/server.ts
@@ -6,6 +6,8 @@ import { RequestWithUser } from 'common/typings'
 import compression from 'compression'
 import cookieParser from 'cookie-parser'
 import session from 'cookie-session'
+import { CMSService, ContentRouter } from 'core/cms'
+import { ExternalAuthConfig, ConfigProvider } from 'core/config'
 import { ConverseRouter, ConverseService } from 'core/converse'
 import { LogsService, LogsRepository } from 'core/logger'
 import cors from 'cors'
@@ -29,13 +31,9 @@ import portFinder from 'portfinder'
 import { URL } from 'url'
 import yn from 'yn'
 
-import { ExternalAuthConfig } from './config/botpress.config'
-import { ConfigProvider } from './config/config-loader'
-
 import { ModuleLoader } from './module-loader'
 import { TelemetryRepository } from './repositories/telemetry'
 import { AdminRouter, AuthRouter, BotsRouter, MediaRouter, ModulesRouter } from './routers'
-import { ContentRouter } from './routers/bots/content'
 import { HintsRouter } from './routers/bots/hints'
 import { NLURouter } from './routers/bots/nlu'
 import { isDisabled } from './routers/conditionalMiddleware'
@@ -52,7 +50,6 @@ import { AuthStrategies } from './services/auth-strategies'
 import AuthService, { EXTERNAL_AUTH_HEADER, SERVER_USER, TOKEN_AUDIENCE } from './services/auth/auth-service'
 import { generateUserToken } from './services/auth/util'
 import { BotService } from './services/bot-service'
-import { CMSService } from './services/cms'
 
 import { FlowService } from './services/dialog/flow/service'
 import { SkillService } from './services/dialog/skill/service'
