+++ src/bp/core/services/auth/auth-service.ts
@@ -2,9 +2,9 @@ import { Logger, StrategyUser } from 'botpress/sdk'
 import { JWT_COOKIE_NAME } from 'common/auth'
 import { StrategyUsersRepository } from 'core/collaborators/repositories/strategy_users'
 import { StrategyUserTable } from 'core/collaborators/tables/strategy_users'
-import { AuthStrategy } from 'core/config/botpress.config'
-import { ConfigProvider } from 'core/config/config-loader'
+import { AuthStrategy, ConfigProvider } from 'core/config'
 import Database from 'core/database'
+import { SessionIdFactory } from 'core/dialog'
 import { EventEngine } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { getMessageSignature } from 'core/misc/security'
@@ -21,7 +21,6 @@ import ms from 'ms'
 import { AuthPayload, AuthStrategyConfig, ChatUserAuth, TokenUser, TokenResponse } from '../../../common/typings'
 import { Resource } from '../../misc/resources'
 import { TYPES } from '../../types'
-import { SessionIdFactory } from '../dialog/session/id-factory'
 import { JobService } from '../job-service'
 import { WorkspaceService } from '../workspace-service'
