+++ src/bp/core/services/bot-service.ts
@@ -3,7 +3,8 @@ import cluster from 'cluster'
 import { BotHealth, ServerHealth } from 'common/typings'
 import { BotCreationSchema, BotEditSchema, isValidBotId } from 'common/validation'
 import { createForGlobalHooks } from 'core/api'
-import { ConfigProvider } from 'core/config/config-loader'
+import { CMSService } from 'core/cms'
+import { ConfigProvider } from 'core/config'
 import { PersistedConsoleLogger } from 'core/logger'
 import { IDisposable } from 'core/misc/disposable'
 import { listDir } from 'core/misc/list-dir'
@@ -29,7 +30,6 @@ import { VError } from 'verror'
 import { extractArchive } from '../misc/archive'
 
 import { InvalidOperationError } from './auth/errors'
-import { CMSService } from './cms'
 import { ReplaceContent } from './ghost'
 import { FileContent, GhostService } from './ghost/service'
 import { Hooks, HookService } from './hook/hook-service'
