+++ src/bp/core/services/services.inversify.ts
@@ -1,5 +1,6 @@
 import { IO } from 'botpress/sdk'
 import LicensingService from 'common/licensing-service'
+import { CMSService } from 'core/cms'
 import { ConverseService } from 'core/converse'
 import { EventEngine, Queue, MemoryQueue } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
@@ -17,7 +18,6 @@ import { AuthStrategies, CEAuthStrategies } from './auth-strategies'
 import AuthService from './auth/auth-service'
 import { BotMonitoringService } from './bot-monitoring-service'
 import { BotService } from './bot-service'
-import { CMSService } from './cms'
 import { SkillService } from './dialog/skill/service'
 import { GhostContainerModule } from './ghost/ghost.inversify'
 import { HintsService } from './hints'
