+++ src/bp/core/services/telemetry/legacy-stats.ts
@@ -1,6 +1,7 @@
 import LicensingService from 'common/licensing-service'
 import { machineUUID } from 'common/stats'
-import { ConfigProvider } from 'core/config/config-loader'
+import { CMSService } from 'core/cms'
+import { ConfigProvider } from 'core/config'
 import Database from 'core/database'
 import { UserRepository } from 'core/repositories'
 import { TelemetryRepository } from 'core/repositories/telemetry'
@@ -14,7 +15,6 @@ import yn from 'yn'
 import { GhostService } from '..'
 import AuthService from '../auth/auth-service'
 import { BotService } from '../bot-service'
-import { CMSService } from '../cms'
 import { JobService } from '../job-service'
 import { WorkspaceService } from '../workspace-service'
